moved reporttsdiagnostics into utils
