+++ packages/modular-scripts/src/buildPackage/makeTypings.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import * as ts from 'typescript';
 import * as fse from 'fs-extra';
 import { getLogger } from './getLogger';
-import { reportTSDiagnostics } from './reportTSDiagnostics';
+import { reportTSDiagnostics } from '../utils/reportTSDiagnostics';
 import getPackageMetadata from '../utils/getPackageMetadata';
 
 const outputDirectory = 'dist';
